HomeScreen Update
Added gradient rectangular buttons to the home screen and put them into
a relative layout to each other.
Added NYAS logo and the empty speech bubble to the top of the screen and
added them to the relative layout.
All colours of the gradient files are held within the values/colours.xml
file.