Home screen update 2
Added a text view to be laid on top of the speech bubble. They are
wrapped up by a relative layout.
A random message is picked when opening the home screen.
Added text to the buttons. Top left button loads the kids zone state and
the top right button the default layout of the page.
String arrays within the Strings.xml file hold the button strings.
Constant / final values are used for the switch case that updates the
state of the page.