HomePage Back Button Update
Added back button functionality:
If the user is in the default state of the activity it will still quit
the app, but changing the state will now be remembered in the
"previousState" integer. Pressing the back button will load the previous
state of the screen.
Currently this means pressing it twice will take you back again to where
you came from.
This can be changed if it was remembered in an ever-growing integer list
that grows and shrinks when pressing the back button.