Appointments Update #1
Appointments can now be saved after you pick a date from the calendar
and select a time after pressing the make appointment button.
They will be saved in a file as strings. The date and time are comma
separated as well as have a comma at the end.
New appointment screen that shows all appointments in a listview.
Updated button gradients to more suitable and eye-pleasing colours.