Profile Update
Added a user profile that can be accessed and edited after entering the
pin. It uses a separator like the other parts of the app that save data.
Since the user can control the entry the app makes sure that when you
try to enter anything that contains the separator char to tell the user
not to use special characters and does not save the corruptive data.