Email Verification
Added email verification to profile activity. This now checks that the
email contains "@", "." and is at least 10 characters long.
You can still just type in anything that contains @, . and make it at
least 10 characters long and it will be accepted. This logic could be
expanded to further verify the users email address.