Information Screen Update #3
Added information screen for carers and an option to send an email from
the "Become a carer" button.
Also added two new features to the options screen:
Delete my profile and delete all appointments. They do exactly what you
would expect. The user is warned at the bottom of the screen that these
options are permanent but currently will not be asked if they are sure.
This should be added in the future.